refactoring for dependency injection

+++ src/modules/cars/useCases/ImportCategory/ImportCategoryUseCase.ts
@@ -1,44 +1,50 @@
 import fs from 'fs'
 import { parse } from 'csv-parse'
 import { ICategoriesRepository } from '../../repositories/ICategoriesRepository'
+import { inject, injectable } from 'tsyringe'
 
 interface IImportCategory {
     name: string
     description: string
 }
 
+@injectable()
 class ImportCategoryUseCase {
-  constructor(private categoriesRepository: ICategoriesRepository) {}
+  constructor(
+    @inject('CategoriesRepository')
+    private categoriesRepository: ICategoriesRepository) {}
 
-  loadCategories(file: Express.Multer.File): Promise<IImportCategory[]> {
+  loadCategories(file: Express.Multer.File | undefined): Promise<IImportCategory[]> {
     return new Promise((resolve, reject) => {
-      const stream = fs.createReadStream(file.path)
-      const categories: IImportCategory[] = []
+      if (file) {
+        const stream = fs.createReadStream(file.path)
+        const categories: IImportCategory[] = []
 
-      const parseFile = parse()
+        const parseFile = parse()
 
-      stream.pipe(parseFile)
+        stream.pipe(parseFile)
 
-      parseFile.on('data', async (line) => {
-        const [ name, description ]= line
-        categories.push({ name, description })
-      })
-        .on('end', () => {
-          fs.promises.unlink(file.path)
-          resolve(categories)})
-        .on('error', (err) => {reject(err)})
+        parseFile.on('data', async (line) => {
+          const [ name, description ]= line
+          categories.push({ name, description })
+        })
+          .on('end', () => {
+            fs.promises.unlink(file.path)
+            resolve(categories)})
+          .on('error', (err) => {reject(err)})
+      }
     })
   }
 
-  async execute(file: Express.Multer.File): Promise<void> {
+  async execute(file: Express.Multer.File | undefined): Promise<void> {
     const categories = await this.loadCategories(file)
         
     categories.map(async category => {
       const { name, description } = category
 
-      const existCategory = this.categoriesRepository.findByName(name)
+      const existCategory = await this.categoriesRepository.findByName(name)
       if (!existCategory) {
-        this.categoriesRepository.create({ name, description })
+        await this.categoriesRepository.create({ name, description })
       }
     })
   }
